Refactor frontend into separate module
