+++ askama.toml
@@ -1,0 +1,3 @@
+[general]
+# Directories to search for templates, relative to the crate root.
+dirs = ["src/frontend/html"]
