+++ src/frontend/js/colored-buttons.js
